pyproject.toml
[project]
name = "ai-agent-boot-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
